working?

+++ maxheap.c
@@ -1,10 +1,10 @@
-#include "main.c"
+#include "headers.h"
 
 //code used to implement this heap was influenced by robin thomas. The code for his heap can be found here https://github.com/robin-thomas/max-heap/blob/master/maxHeap.c
 
 //define node structure
 typedef struct node {
-    int v ;
+    int64_t v ;
 } node ;
